Add test setup/environment APIs
