+++ pkg/new/onit/setup/app.go
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package onit
+package setup
 
-import "github.com/onosproject/onos-test/pkg/new/kubetest"
+import (
+	corev1 "k8s.io/api/core/v1"
+)
 
-// RegisterTests registers a test suite
-func RegisterTests(name string, suite TestSuite) {
-	kubetest.RegisterTests(name, suite)
-}
-
-// RegisterBenchmarks registers a benchmark suite
-func RegisterBenchmarks(name string, suite BenchmarkSuite) {
-	kubetest.RegisterBenchmarks(name, suite)
+// AppSetup is an interface for setting up an application
+type AppSetup interface {
+	Setup
+	Name(name string) AppSetup
+	Image(image string) AppSetup
+	PullPolicy(pullPolicy corev1.PullPolicy) AppSetup
 }
